+++ src/app/components/columns-configuration-popover.js
@@ -110,6 +110,11 @@ export default Component.extend(...mixins, {
   modifiedJsonKey: '',
 
   /**
+   * @type {string}
+   */
+  modifiedJsonQuery: '',
+
+  /**
    * @type {ComputedProperty<string>}
    */
   columnsCount: reads('columnsConfiguration.columnsOrder.length'),
@@ -318,6 +323,11 @@ export default Component.extend(...mixins, {
           this.set(
             'modifiedJsonKey',
             columnInfo.options.jsonKey
+          );
+        } else if (queryType === 'query') {
+          this.set(
+            'modifiedJsonQuery',
+            columnInfo.options.jsonQuery
           );
         }
       }