+++ src/app/locales/en/components/-browser-columns-name/filesystem.js
@@ -1,3 +1,5 @@
+const customJson = 'Custom JSON metadata assigned to the file';
+
 export default {
   name: {
     firstColumnName: 'Files',
@@ -23,8 +25,9 @@ export default {
     posixPermissions: 'Permissions regulating data access rights. It can be expressed either by the standard POSIX permissions system or by the Access Control List (ACL). If there is an ACL specified, the POSIX permissions are ignored.<br>Click on the cell to see more details.',
     xattr: 'The extended attribute value for key: <strong>{{key}}</strong>',
     json: {
-      all: 'Custom JSON metadata assigned to the file (mode: whole document).',
-      key: 'Custom JSON metadata assigned to the file (mode: extract key <strong>"{{key}}"</strong>).',
+      all: `${customJson} (mode: whole document).`,
+      key: `${customJson} (mode: extract key <strong>"{{key}}"</strong>).`,
+      query: `${customJson} (mode: extract JSONata query <strong>"{{query}}"</strong>).`,
     },
     fileId: {
       priv: 'File ID is a unique, global identifier associated with a file/directory. It can be used to access and interact with the file/directory using different interfaces (e.g. REST API), but only by authorized users that have permission to do so.',