+++ src/app/locales/en/components/columns-configuration-popover/column-editor.js
@@ -2,6 +2,11 @@ export default {
   addNewColumnHeader: 'New metadata column',
   columnNameLabel: 'Column label:',
   jsonKeyLabel: 'Top-level key:',
+  jsonQueryLabel: 'Query:',
+  jsonQueryDescription: 'Enter a query using',
+  jsonQueryDescription2: 'syntax to extract specific information from the JSON data.',
+  jsonataLink: 'https://docs.jsonata.org/overview.html',
+  jsonataLinkText: 'JSONata',
   addBtn: 'Create',
   cancelBtn: 'Cancel',
   modifyColumnHeader: 'Modify metadata column',
@@ -10,6 +15,7 @@ export default {
   columnExistsTooltip: 'This column already exists.',
   emptyValueTooltip: 'Requires non-empty inputs.',
   columnLabelExistsTooltip: 'Column with this label already exists.',
+  invalidJsonataExpressionTooltip: 'Invalid JSONata expression: {{error}}',
   xattrKey: {
     label: 'Extended attribute key',
     customValueInputPlaceholder: 'Enter or choose a key...',
@@ -40,6 +46,11 @@ export default {
       key: {
         label: 'Extract key',
       },
+      query: {
+        label: 'Query',
+      },
     },
   },
+  json: 'JSON',
+  jsonQuery: 'JSON query',
 };